+++ src/main/java/io/github/astrapi69/tree/element/LazyTreeElement.java
@@ -22,9 +22,8 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.tree;
+package io.github.astrapi69.tree.element;
 
-import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
@@ -32,44 +31,27 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-import lombok.experimental.FieldDefaults;
-import io.github.astrapi69.tree.api.IChainableTreeNode;
 
 /**
- * The class {@link ChainableTreeNode} can have only one child and one parent
+ * The class {@link LazyTreeElement} represents a lazy tree element that does not know about the
+ * parent or children
  *
  * @param <T>
  *            the generic type of the value
  */
 @Getter
 @Setter
-@EqualsAndHashCode(exclude = { "child" })
-@ToString(exclude = { "child" })
+@EqualsAndHashCode
+@ToString
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder(toBuilder = true)
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class ChainableTreeNode<T> implements IChainableTreeNode<T>
+public class LazyTreeElement<T>
 {
 
-	/** The single child. */
-	IChainableTreeNode<T> child;
-
-	/** The single parent. */
-	IChainableTreeNode<T> parent;
-
 	/** The value. */
-	T value;
+	private T value;
 
-	/**
-	 * Instantiates a new {@link ChainableTreeNode} object.
-	 *
-	 * @param value
-	 *            the value
-	 */
-	public ChainableTreeNode(T value)
-	{
-		this.value = value;
-	}
-
+	/** The flag that indicates if this tree element is a leaf. */
+	private boolean leaf;
 }