+++ src/test/java/io/github/astrapi69/tree/TreeNodeTest.java
@@ -44,6 +44,7 @@ import org.testng.annotations.Test;
 import io.github.astrapi69.AbstractTestCase;
 import io.github.astrapi69.evaluate.object.evaluators.EqualsHashCodeAndToStringEvaluator;
 import io.github.astrapi69.tree.api.ITreeNode;
+import io.github.astrapi69.tree.element.TreeElement;
 
 /**
  * The unit test class for the class {@link TreeNode}