+++ src/test/java/io/github/astrapi69/tree/element/TreeElementNodeTest.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.tree;
+package io.github.astrapi69.tree.element;
 
 import static org.testng.AssertJUnit.assertNotNull;
 
@@ -31,6 +31,8 @@ import org.meanbean.test.BeanTester;
 import org.meanbean.test.Configuration;
 import org.meanbean.test.ConfigurationBuilder;
 import org.testng.annotations.Test;
+
+import io.github.astrapi69.tree.TreeNode;
 
 /**
  * The unit test class for the class {@link TreeElementNode}