+++ src/test/java/io/github/astrapi69/tree/visitor/TraverseSimpleTreeNodeVisitor.java
@@ -1,10 +1,11 @@
-package io.github.astrapi69.tree;
+package io.github.astrapi69.tree.visitor;
 
 import java.util.LinkedHashSet;
 import java.util.Set;
 
 import lombok.Getter;
 import io.github.astrapi69.design.pattern.visitor.Visitor;
+import io.github.astrapi69.tree.SimpleTreeNode;
 
 public class TraverseSimpleTreeNodeVisitor<T> implements Visitor<SimpleTreeNode<T>>
 {